Extract IP manipulation logic to a separate component
